bot_v2/requirements.txt
keyboard
pygetwindow
pyautogui
easyocr
Pillow
